third_party/blink/web_tests/http/tests/xmlhttprequest/sync-xhr-revalidate-after-async-xhr-expected.txt

Sync XHR should not turn async due to revalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr2SendIsReturned is false
PASS xhr1.responseText is not xhr2.responseText
PASS successfullyParsed is true

TEST COMPLETE
